Move code into Parquet
